some fixes for some issues
