+++ components/fileUploader/index.js
@@ -2,6 +2,9 @@ import React, { useState, useEffect } from 'react';
 // import ReactFilestack from 'filestack-react';
 import PropTypes from 'prop-types';
 import dynamic from 'next/dynamic';
+import {
+  isClientSide,
+} from '../../utils/utilities';
 import handleError from '../../utils/errorHandler';
 
 const UploadedFile = dynamic(() => import('./uploadedFile'));
@@ -16,7 +19,7 @@ const FileUploader = ({
   apiKey, buttonText, requiredFile, inputName,
 }) => {
   const [uploadedFiles, setUploadedFiles] = useState([]);
-
+  const isClient = isClientSide();
   const fetchFileStackLibrary = async () => {
     // eslint-disable-next-line global-require
     ReactFilestack = await require('filestack-react');
@@ -41,39 +44,42 @@ const FileUploader = ({
     setUploadedFiles(uploadedFiles.concat(fileList));
   };
 
-  return typeof window !== 'undefined' ? (
+  return (
     <>
-      <ReactFilestack
-        apikey={apiKey}
-        onSuccess={onSuccess}
-        onError={(error) => { handleError(error, `error ${error}`); }}
-        render={({ onPick }) => (
-          <div className="relative">
-            {/* eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions */}
-            <label
-              onClick={onPick}
-              onKeyPress={onPick}
-              className="document-item--button"
-              htmlFor={`uploadFile-${apiKey}`}
-            >
-              {buttonText}
-            </label>
-            {requiredFile && uploadedFiles.length === 0 && (
-              <input
-                id={`uploadFile-${apiKey}`}
-                type="file"
-                style={hiddenStyle}
-                required={requiredFile}
-                name="File upload"
-                className="required-document-item--button"
-              />
-            )}
-          </div>
-        )}
-      />
+      {isClient
+          && (
+            <ReactFilestack
+              apikey={apiKey}
+              onSuccess={onSuccess}
+              onError={(error) => { handleError(error, `error ${error}`); }}
+              render={({ onPick }) => (
+                <div className="relative">
+                  {/* eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions */}
+                  <label
+                    onClick={onPick}
+                    onKeyPress={onPick}
+                    className="document-item--button"
+                    htmlFor={`uploadFile-${apiKey}`}
+                  >
+                    {buttonText}
+                  </label>
+                  {requiredFile && uploadedFiles.length === 0 && (
+                    <input
+                      id={`uploadFile-${apiKey}`}
+                      type="file"
+                      style={hiddenStyle}
+                      required={requiredFile}
+                      name="File upload"
+                      className="required-document-item--button"
+                    />
+                  )}
+                </div>
+              )}
+            />
+          )}
       {uploadedFiles}
     </>
-  ) : null;
+  );
 };
 
 FileUploader.propTypes = {
