+++ components/universitySelector/index.js
@@ -1,23 +1,34 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { loadUniversities } from '../../services/positions';
 
-const UniversitySelector = ({
-  placeholder, classNames, containerClassNames, onChangeFunction, universities,
-}) => (
-  <div className={containerClassNames}>
-    <select
-      placeholder={placeholder}
-      name="uni-campus-attending"
-      className={classNames}
-      onChange={onChangeFunction}
-      defaultValue=""
-      required
-    >
-      <option value="" disabled>
-        {placeholder}
-      </option>
-      {universities.length
+// getInitialProps can only be used inside the pages directory
+const UniversitySelector = (props) => {
+  const {
+    placeholder, classNames, containerClassNames, onChangeFunction,
+  } = props;
+  const [universities, setUniversities] = useState([]);
+
+  useEffect(() => {
+    loadUniversities().then((unis) => { // this is not working.
+      setUniversities(unis);
+    });
+  }, [universities]);
+
+  return (
+    <div className={containerClassNames}>
+      <select
+        placeholder={placeholder}
+        name="uni-campus-attending"
+        className={classNames}
+        onChange={onChangeFunction}
+        defaultValue=""
+        required
+      >
+        <option value="" disabled>
+          {placeholder}
+        </option>
+        {universities.length
         && universities
           .filter((university) => university.text && university.value)
           .map((university) => (
@@ -25,13 +36,9 @@ const UniversitySelector = ({
               {university.text}
             </option>
           ))}
-    </select>
-  </div>
-);
-
-UniversitySelector.getInitialProps = async () => {
-  const universities = await loadUniversities();
-  return { universities };
+      </select>
+    </div>
+  );
 };
 
 UniversitySelector.propTypes = {
@@ -39,7 +46,6 @@ UniversitySelector.propTypes = {
   classNames: PropTypes.string,
   containerClassNames: PropTypes.string,
   onChangeFunction: PropTypes.func,
-  universities: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
 };
 
 UniversitySelector.defaultProps = {
