--- pages/positions/[positionId]/[jobCategory]/positionsEntry.scss
@@ -1,7 +1,0 @@
-@import "../../../../styles/base/variables";
-
-.job-hero {
-  background:
-    url($cdnRoot+"assets/images/banner/Partner@2x-min.jpg") no-repeat
-    bottom center / cover #050301;
-}
